feat(server): carry the bestiary from the row to the runtime and back through the ledger (FUN-113)
Coluna nulável bestiary em character (0005), ticket com o mapa validado, extrato absoluto e
fusão pelo maior no ledger — o padrão das skills. O host manda `bestiary` no attach e quando a
soma muda; o catálogo leva os monstros e os marcos. O extrato do snapshot que este nó não
sabe retomar passa a levar skills e Bestiário: a XP era creditada e o abate 9 999 voltava
a ser o do banco. O bônus por marco é inteiro no schema, porque a conta em inteiro depende disso.

+++ packages/content/src/schemas.ts
@@ -498,8 +498,15 @@ export const bestiarySchema = z.object({
       return;
     }
   }),
-  /** Quanto cada marco acrescenta à XP PvE, em pontos percentuais. §18.3: 1. */
-  xpBonusPercentPerMilestone: z.number().nonnegative(),
+  /**
+   * Quanto cada marco acrescenta à XP PvE, em pontos percentuais. §18.3: 1.
+   *
+   * INTEIRO, e é regra de forma que sustenta uma de conta: `Bestiary.applyXpBonus` faz
+   * `floor(xp × (100 + p × marcos) / 100)` em inteiro, e a garantia de que o `floor` acerta
+   * depende de `p × marcos` ser inteiro. Meio ponto percentual voltaria a pôr resíduo de
+   * ponto flutuante na frente do arredondamento — a armadilha que a conta em inteiro evita.
+   */
+  xpBonusPercentPerMilestone: z.number().int().nonnegative(),
   _open: z.string().optional(),
 });
